[SPARK-25299] SPARK-25299 upstream updates (#605)
* Bring implementation into closer alignment with upstream.

Step to ease merge conflict resolution and build failure problems when we pull in changes from upstream.

* Cherry-pick BypassMergeSortShuffleWriter changes and shuffle writer API changes

* [SPARK-28607][CORE][SHUFFLE] Don't store partition lengths twice

The shuffle writer API introduced in SPARK-28209 has a flaw that leads to a memory usage regression - we ended up tracking the partition lengths in two places. Here, we modify the API slightly to avoid redundant tracking. The implementation of the shuffle writer plugin is now responsible for tracking the lengths of partitions, and propagating this back up to the higher shuffle writer as part of the commitAllPartitions API.

Existing unit tests.

Closes #25341 from mccheah/dont-redundantly-store-part-lengths.

Authored-by: mcheah <mcheah@palantir.com>
Signed-off-by: Marcelo Vanzin <vanzin@cloudera.com>

* [SPARK-28571][CORE][SHUFFLE] Use the shuffle writer plugin for the SortShuffleWriter

Use the shuffle writer APIs introduced in SPARK-28209 in the sort shuffle writer.

Existing unit tests were changed to use the plugin instead, and they used the local disk version to ensure that there were no regressions.

Closes #25342 from mccheah/shuffle-writer-refactor-sort-shuffle-writer.

Lead-authored-by: mcheah <mcheah@palantir.com>
Co-authored-by: mccheah <mcheah@palantir.com>
Signed-off-by: Marcelo Vanzin <vanzin@cloudera.com>

* [SPARK-28570][CORE][SHUFFLE] Make UnsafeShuffleWriter use the new API.

* Resolve build issues and remaining semantic conflicts

* More build fixes

* More build fixes

* Attempt to fix build

* More build fixes

* [SPARK-29072] Put back usage of TimeTrackingOutputStream for UnsafeShuffleWriter and ShufflePartitionPairsWriter.

* Address comments

* Import ordering

* Fix stream reference