[SPARK-25299] Put SPARK-25299 changes into master (#606)
## What changes were proposed in this pull request?

Pushes SPARK-25299 changes into our master branch. Some PRs have merged upstream, and we intend to ship the rest of them moving forward.

## How was this patch tested?

We've been iteratively testing this in unit tests as well as with TPC-DS. We're using the API changes internally. We intend to test these changes on test stacks moving forward.