[SPARK-25299] Add shuffle map output un-registration hooks upon fetch failure (#609)
We realized that there's complexity with whether or not map outputs should be unregistered, and thus should be recomputed.

Previously, we were using a boolean, then, a three-way switch. But these modes do not capture a lot of intricacies with how this should work - in particular, for our async upload proof of concept, we end up unregistering all the map outputs written by an executor, despite the fact that this would invalidate and re-write all the map outputs that were persisted to the remote storage system.